fix player  ordering
