+++ src/client/src/App.jsx
@@ -10,30 +10,31 @@ const socket = io.connect(endpoint);
 
 const App = () => {
   const [gameState, setGameState] = useState();
+  const [seat, setSeat] = useState(0);
 
   socket.on('connect', () => {
     console.log('in app.js ' + socket.id + ' successfully connected');
   });
 
   socket.on('stateUpdated', (data) => {
-    console.log('in app.js data', data);
     setGameState(data);
+  });
+
+  socket.on('joinGameSuccess', (data) => {
+    console.log('in app.js join game success', data);
+    setSeat(data.seat);
   });
 
   return (
     <BrowserRouter>
       <div>
         <Switch>
-          <Route
-            exact
-            path="/"
-            render={(props) => <Home {...props} {...gameState} socket={socket} />}
-          />
+          <Route exact path="/" render={(props) => <Home {...props} socket={socket} />} />
           <Route
             path="/game/:id"
-            render={(props) => <GamePath {...props} {...gameState} socket={socket} />}
+            render={(props) => <GamePath {...props} {...gameState} seat={seat} socket={socket} />}
           />
-          <Route render={(props) => <NotFound {...props} {...gameState} />} />
+          <Route render={(props) => <NotFound {...props} />} />
         </Switch>
       </div>
     </BrowserRouter>
